--- Insurance-System/src/main/java/org/example/InvestProductListImpl.java
@@ -1,35 +1,0 @@
-package org.example;
-
-import java.util.List;
-
-/**
- * @author eirch
- * @version 1.0
- */
-public class InvestProductListImpl implements InvestProductList {
-
-	private List<InvestProduct> InvestProducts;
-	public InvestProduct m_InvestProduct;
-
-
-	public void finalize() throws Throwable {
-
-	}
-
-	public void add(){
-
-	}
-
-	public void delete(){
-
-	}
-
-	public void read(){
-
-	}
-
-	public void update(){
-
-	}
-
-}
