+++ Insurance-System/src/main/java/org/example/asset/InvestProduct.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.asset;
 
 /**
  * @author eirch
