+++ Insurance-System/src/main/java/org/example/asset/InvestProfitScenario.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.asset;
 
 /**
  * @author USER
