+++ Insurance-System/src/main/java/org/example/asset/InvestResult.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.asset;
 
 /**
  * @author USER
