+++ Insurance-System/src/main/java/org/example/asset/InvestResultListImpl.java
@@ -1,0 +1,36 @@
+package org.example.asset;
+
+/**
+ * @author USER
+ * @version 1.0
+ */
+public class InvestResultListImpl implements InvestResultList {
+
+	private InvestResult investResults;
+	public InvestResult m_InvestResult;
+
+	public InvestResultListImpl(){
+
+	}
+
+	public void finalize() throws Throwable {
+
+	}
+
+	public void add(){
+
+	}
+
+	public void delete(){
+
+	}
+
+	public void read(){
+
+	}
+
+	public void update(){
+
+	}
+
+}
