+++ Insurance-System/src/main/java/org/example/asset/InvestScenario.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.asset;
 
 /**
  * @author USER
