+++ Insurance-System/src/main/java/org/example/asset/InvestScenarioListImpl.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.asset;
 
 import java.util.List;
