+++ Insurance-System/src/main/java/org/example/asset/LoanHistory.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.asset;
 
 public class LoanHistory {
 }
