+++ Insurance-System/src/main/java/org/example/complaint/Complaint.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.complaint;
 
 import org.example.user.Customer;
