+++ Insurance-System/src/main/java/org/example/complaint/ComplaintCategory.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.complaint;
 
 public enum ComplaintCategory {
 }
