+++ Insurance-System/src/main/java/org/example/complaint/ComplaintEvaluation.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.complaint;
 
 /**
  * @author USER
