+++ Insurance-System/src/main/java/org/example/complaint/ComplaintStatus.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.complaint;
 /**
  * @author USER
  * @version 1.0
